requirements.txt
pymeshlab==2022.2.post2
numpy==1.24.1
pandas==1.5.2
seaborn==0.12.2
jupytext==1.14.4
PyYAML==6.0
